requirements.txt
streamlit
google-generativeai
langchain-google-genai
python-dotenv